A clever idea, but not sure how I want to implement it...
Currently does not compile- but I had the idea to make the video bus
cycles into a state machine.  Whether to make it one big enum of each
cycle across the entire line, or a smaller enum of state types sith some
sort of container class to hold a cycle type and number, then each
variant can have its own collection of these objects with different
numbers of idle cycles...  Difficulty is some cycles have to look ahead
at what is coming to for example raise the BA signal 3 cycles before
cycle stealing needs to happen...  committing this half baked idea so I
don't lose it to the minutiae of implementing it.